ticket-kafka/ticket-fe/requirements.txt
flask==0.10.1
redis==3.2.1
kafka-python==1.4.6
